fourth_semester/Python/Pycasso/requirements.txt
numpy==1.13.0+mkl
Keras==2.0.4
scipy==0.19.0
tensorflow_gpu==1.0.1
matplotlib==2.0.2
Pillow==4.1.1
tensorflow==1.2.0rc2
